generating destination and labels in a list
Got the output into a format that’s ready to be concatenated together
into a string for tiny choice. We make a list of spiral points
generated using different distributions and then loop over them to
perform all the necessary checks and output the text label and name of
the destination.